Fixed uncaught ImmediateHttpResponse when adapter is subclassed.
views.py assumed that the adapter would always be OTPAdapter, and calls login()
on its parent to get DefaultAccountAdapter.login().

This fails when the adapter is subclassed from OTPAdapter (for other reasons, in my case that I needed to override a method unrelated to 2FA), because
OTPAdapter.login() gets called by TwoFactorAuthenticate.form_valid(),
generating an inadvertent and uncaught ImmediateHttpResponse.

The solution is to add a method to OTPAdapter that calls
DefaultAccountAdapter.login() and have TwoFactorAuthenticate.form_valid() call
it.